Move ChangedPixelUnitData to restorer package
